adding pip installation ability
