+++ .gitignore
@@ -2,3 +2,4 @@
 *.npy
 *.npz
 *.fam
+.idea*
