--- Matrices/Epistasis.py
@@ -1,2 +1,0 @@
-def pairwise_epistasis(ibd):
-    return ibd.multiply(ibd)
