+++ SciLMM/Matrices/Epistasis.py
@@ -1,0 +1,2 @@
+def pairwise_epistasis(ibd):
+    return ibd.multiply(ibd)
