+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(name='scilmm',
+      version='0.1',
+      description='Sparse Cholesky factorIzation Linear Mixed model',
+      url='http://github.com/kalkairis/SciLMM.git',
+      author='TalShor',
+      author_email='talihow@gmail.com',
+      license='GNU',
+      packages=['scilmm'],
+      zip_safe=False)
